COM-20 renaming hooks and actions
